+++ lib/src/interactive/image_interactive_media.dart
@@ -6,8 +6,7 @@ import '../media_provider.dart';
 /// Displays zoomable images
 class ImageInteractiveMedia extends StatelessWidget {
   final MediaProvider mediaProvider;
-  const ImageInteractiveMedia({Key? key, required this.mediaProvider})
-      : super(key: key);
+  const ImageInteractiveMedia({super.key, required this.mediaProvider});
 
   @override
   Widget build(BuildContext context) {