+++ lib/src/interactive/text_interactive_media.dart
@@ -7,8 +7,7 @@ import 'package:flutter/material.dart';
 class TextInteractiveMedia extends StatelessWidget {
   final MediaProvider mediaProvider;
 
-  const TextInteractiveMedia({Key? key, required this.mediaProvider})
-      : super(key: key);
+  const TextInteractiveMedia({super.key, required this.mediaProvider});
 
   @override
   Widget build(BuildContext context) {